# golang/product-api/swagger.yaml
definitions:
    Product:
        description: Product defines the structure for an API product
        properties:
            description:
                type: string
                x-go-name: Description
            id:
                format: int64
                type: integer
                x-go-name: ID
            name:
                type: string
                x-go-name: Name
            price:
                format: float
                type: number
                x-go-name: Price
            sku:
                type: string
                x-go-name: SKU
        type: object
        x-go-package: github.com/nicholasjackson/building-microservices-youtube/product-api/data
info: {}
paths:
    /products:
        get:
            description: Return a list of products from the data store
            operationId: listProducts
            responses:
                "200":
                    $ref: '#/responses/productsResponse'
                "404":
                    $ref: '#/responses/errorResponse'
            tags:
                - products
    /products/{id}:
        delete:
            description: Delete a product from the database
            operationId: deleteProduct
            parameters:
                - description: The id of the product to delete from database
                  format: int64
                  in: path
                  name: id
                  required: true
                  type: integer
                  x-go-name: ID
            responses:
                "201":
                    $ref: '#/responses/noContentResponse'
                "404":
                    $ref: '#/responses/errorResponse'
                "500":
                    $ref: '#/responses/errorResponse'
            tags:
                - products
responses:
    productsResponse:
        description: A list of products returned in the response
        schema:
            items:
                $ref: '#/definitions/Product'
            type: array
swagger: "2.0"